[Architecture] temporary renaming
