+++ test/pa/finder_test.py
@@ -1,12 +1,12 @@
 import unittest
 import pytest
-from ...Modules.PhotoAnalyzer import image_processor
+from ...Modules.pa import finder
 from ...Modules.image_data import ImageData
 
 class TestMethods(unittest.TestCase):
     @classmethod
     def setup_class(cls):
-        global its
+        global f
 
         text = """
             caFE PARVIS
@@ -37,21 +37,19 @@ class TestMethods(unittest.TestCase):
             $ 291.20
             """
 
-        its = image_processor.ImageTextSearch(text)
-        image_processor.LOCAL_DEBUG = True
+        f = finder.Finder(text)
 
-    def test_analyze(self):
-        attrs = {
-            "date": "07/01/2017",
-            "time": "20:48:14",
-            "address": None,
-            "total_amount": "$2017.21",
-            "description": None,
-        }
+    def test_find_date(self):
+        assert "07/01/2017" == f.find_date()
 
-        image_data = its.analyze()
+    def test_find_time(self):
+        assert "20:48:14" == f.find_time()
 
-        assert isinstance(image_data, ImageData)
+    def test_find_address(self):
+        assert None == f.find_address()
 
-        for attr, attr_val in attrs.iteritems():
-            assert(attr_val == image_data.__dict__[attr])
+    def test_find_total_amount(self):
+        assert "$2017.21" == f.find_total_amount()
+
+    def test_description(self):
+        assert None == f.find_description()
